+++ CHANGELOG.md
@@ -1,13 +1,21 @@
 # Changelog
-## Unrelease
+## v0.2.5 (PRERELEASE)
 * Added more error handling
 * Some more constraints
 * Removed extra cells from follower and following list
 * Added acitvity indicator
 * Improved error handling
-* Added functioning json post request sender.
 * Signing up now works through the app
 * Fixed the default avatar
+* Fixed api url
+* Added api fetching (10 second interval)
+* Changed SwiftKeychainWrapper to Valet (Should fix the keychain issue)
+* Fixed bug where timer would continue when user would log out.
+* Added "nothing here" on follower list and following list when a user doesn't have any followers or any one they are following.
+* Added account settings view controller (Doesn't actually contact the api. Just placeholder for now.)
+* Moved Nuke dependency from Swift Package Manger to Cocoapods.
+* Gallery for record page works now (but doesn't send any file anywhere)
+* Added bio to channel page.
 ## v0.2 (PRERELEASE)
 * Some design changes
 * Added most channel attrbitues that update (except followers and following)